requirements.txt
Django==5.1.5
djangorestframework==3.15.2
Pillow==10.4.0
python-decouple==3.8